requirements_dev.txt
django>=1.8,<1.9
djangorestframework>=3.2,<3.3
psycopg2
